[Joysticks] Refactor button mapping into new folder
